requirements.txt
anyascii==0.2.0
asgiref==3.4.0
beautifulsoup4==4.9.3
boto3==1.17.102
botocore==1.20.102
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.4
django-activeurl==0.2.0
django-appconf==1.0.4
django-classy-tags==2.0.0
django-cogwheels==0.3
django-filter==2.4.0
django-modelcluster==5.1
django-storages==1.11.1
django-taggit==1.4.0
django-treebeard==4.5.1
djangorestframework==3.12.4
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==2.10
jmespath==0.10.0
l18n==2020.6.1
lxml==4.6.3
openpyxl==3.0.7
Pillow==8.2.0
psycopg2==2.9.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
telepath==0.2
urllib3==1.26.6
wagtail==2.13.2
wagtail-color-panel==1.2.0
wagtailmenus==3.0.2
wagtailuiplus==1.3.14
webencodings==0.5.1
whitenoise==5.2.0
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.4.3
xlwt==1.3.0
